requirements.txt
beautifulsoup4==4.11.1
pandas==1.5.2
python_dateutil==2.8.2
requests==2.28.1
